python/a2a-prompt-runner-py/pyproject.toml
[project]
name = "a2a-prompt-runner-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "a2a-sdk>=0.3.11,<0.4.0",
    "azure-core>=1.36.0,<2.0",
    "azure-identity>=1.25.1,<2.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b44,<2.0",
    "azure-storage-blob>=12.27.1,<13.0",
    "lagom>=2.7.7,<3.0.0",
    "opentelemetry-api>=1.38.0,<2.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.38.0,<2.0",
    "opentelemetry-instrumentation-requests>=0.59b0",
    "opentelemetry-instrumentation==0.59b0",
    "pydantic>=2.12.4,<3.0",
]

[project.optional-dependencies]
cli = [
    "pyright>=1.1.407,<2.0.0",
    "ruff>=0.14.4,<0.15.0",
]


[tool.pyright]
venvPath = "."
venv = ".venv"
defineConstant = { DEBUG = true }
typeCheckingMode = "strict"
include = ["src"]
strict = ["src"]
exclude = [
    ".venv",
    "typings",
    "**/__pycache__",
    "**/test*.py",
    "**/*.as-is.py"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/a2a_prompt_runner",
]
